{{ cookiecutter.project_slug }}/requirements/base.txt
pytz
Pillow
django
django-environ
django-anymail[mailgun]
django-s3-storage
sentry-sdk
django-extensions
celery
pytest
pytest-django
pytest-xdist
whitenoise
gunicorn
psycopg2
factory-boy
ipython
argon2-cffi
{%- if cookiecutter.rest_framework == 'y' %}
djangorestframework
django-cors-headers
drf-yasg
{%- endif %}
